requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sklearn==0.0.post7
threadpoolctl==3.2.0
tzdata==2023.3
Werkzeug==2.3.6
